HTLC set min timeout (#364)
