+++ test/functional/feature_bitcoin_htlc.py
@@ -50,6 +50,13 @@ class BitcoinHTLCTests(DefiTestFramework):
             errorString = e.error['message']
             assert("Invalid block denominated relative timeout" in errorString)
 
+        # Try annd create a HTLC script below min blocks
+        try:
+            self.nodes[0].spv_createhtlc("0224e7de2f3a9d4cdc4fdc14601c75176287297c212aae9091404956955f1aea86", "035fb3eadde611a39036e61d4c8288d1b896f2c94cee49e60a3d1c02236f4be490", "8", seed_hash)
+        except JSONRPCException as e:
+            errorString = e.error['message']
+            assert("Timeout below minimum of" in errorString)
+
         # Try annd create a HTLC script with incorrect pubkey
         try:
             self.nodes[0].spv_createhtlc("0224e7de2f3a9d4cdc4fdc14601c75176287297c212aae9091404956955f1aea", "035fb3eadde611a39036e61d4c8288d1b896f2c94cee49e60a3d1c02236f4be490", "10", seed_hash)
